added authentication

+++ EduNote.API/Controllers/QuestionsController.cs
@@ -25,7 +25,7 @@ namespace EduNote.API.Controllers
         }
 
         [HttpGet("{id}")]
-        public override IActionResult Get(int id)
+        public override IActionResult Get(long id)
         {
             try
             {
